requirements.txt
Django==3.0.3
django-crispy-forms==1.8.1
djangorestframework==3.11.0
